requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.7
Django==5.0.6
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-phonenumber-field==7.3.0
django-tinymce==4.0.0
httplib2==0.22.0
idna==3.7
phonenumbers==8.13.36
pillow==10.3.0
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-decouple==3.8
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
win32-setctime==1.1.0
fontawesomefree==6.5.1